Created a script to delete day old capsule
A command was written to call the function to delete day old models
because django needs to be active to access model data. django-cleanup
was added because it automatically deletes the files of models
instances. This project is being set to production mode. A developmental
mode will be created in a branch.