Added feature to display activate capsules
SnapCapsule class added class variable of when the capsule should be
deleted, so views.py can filter capsule for active capsules. The only
capsule recieved from a Get Request are are schedule to be deleted in
future by less than a day. Settings base was change to reflect the
current new base directory from settings being change to folder.